searching and sorting solved

+++ BitTracking/conf/evolutions/default/1.sql
@@ -22,8 +22,8 @@ create table package (
   tracking_num              varchar(255),
   destination               varchar(255),
   status                    integer,
-  weight                    double,
-  price                     double,
+  weight                    double(10,2),
+  price                     double(10,2),
   constraint ck_package_status check (status in ('1','2','4','3')),
   constraint pk_package primary key (id))
 ;
